create `Page`, `View` and `PageElement` objects(check details).
This update adds a way to define things on a webpage in order to make
it easier to write a controller, or the automation in general.  It is
not complete, however it is useful enough to warrant adding into the
main branch at this point.

The `PageElement` object is essentially just a definition of an element
that could be on a webpage.  It doesn’t have to exist at the time of
instantiation, however when an instance of `PageElement` has it’s
`__call__’ method invoked, it will attempt to retrieve the web element
using it’s definition, and a selenium WebDriver instance that was
provided on instantiation.

The `View` object is used almost like a container for all of the
`PageElement` objects, or anything else that has been set up in the
`View`.  A view takes a JSON object that defines what the web page
layout looks like and sets it’s own attributes based on the JSON.  It
also instantiates `PageElement`s if needed.

The `Page` and `View` object are almost exactly the same.  The `Page`
object is a light wrapper around the `View` object.  It allows a bit of
reflection to be performed in order to decide if it needs to invoke the
`__call__` method on  a `PageElement` when it is accessed as an
attribute.  It also allows the `PageElement`s to be accessed through
the `Page.view` attribute without invoking the `PageElement`’s
`__call__`  method.

Full Example:

    json = “””
    {
      "root": "https://portland.craigslist.org/search/zip",
      "elements": {
        "search_input": {
          "lookup_method": "id",
          "lookup_target": "query"
          },
        "search_button": {
          "lookup_method": "xpath",
          "lookup_target": "//button[@class='searchbtn changed_input
clickme']"
        },
        "item_titles": {
          "multiple": true,
          "lookup_method": "xpath",
          "lookup_target": "//li[@class='result-row']"
        }
      }
    }
    “””
    from selenium.webdriver import Chrome

    driver = Chrome()
    page = Page(driver, json)

    # Check if the driver is on the correct page.
    if not page:
        page.get(page.root)

    # Check if the search input exists and fill it.
    if page.view.search_input.exists():
        page.search_input.send_keys(‘lumber’)

    # Wait for the correct search button to appear.
    while not page.view.search_button.exists():
        sleep(1)

    page.search_button.click()